+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Nuvola SDK Change Log
 * convert-project: Create `template--README.md` & `template--README.md.diff` to help with the update of README.md.
 * new-project: The `integrate.js` template uses Standard JS code style.
 * convert-project: Convert `integrate.js` to Standard JS code style.
+* check-project: Check that `integrate.js` uses Standard JS code style.
 
 4.11.0 - May 8th, 2018
 ----------------------